study(step4): WinningNumbers 추가
