+++ src/test/kotlin/lotto/domain/RankTest.kt
@@ -6,6 +6,7 @@ import lotto.domain.model.Lotto
 import lotto.domain.model.LottoNumber
 import lotto.domain.model.Rank
 import lotto.domain.model.Winning
+import lotto.domain.model.WinningNumbers
 import org.assertj.core.api.AssertionsForClassTypes.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
@@ -83,7 +84,7 @@ internal class RankTest {
     @DisplayName("당첨 된 숫자 개수를 확인한다")
     @Test
     fun matchingCount() {
-        val winning = Winning(Winning.makeLottoNumbers("8, 21, 23, 41, 42, 43"), LottoNumber(1))
+        val winning = Winning(WinningNumbers.makeWinningNumbers("8, 21, 23, 41, 42, 43"), LottoNumber(1))
         val lotto = Lotto(listOf(9, 21, 27, 41, 42, 45).map { LottoNumber(it) })
 
         assertThat(Rank.countMatchNumber(winning, lotto)).isEqualTo(3)
