+++ src/test/kotlin/lotto/domain/model/WinningTest.kt
@@ -12,9 +12,9 @@ internal class WinningTest {
     @DisplayName("입력 된 당첨 번호로 당첨 번호가 생성되어야 한다")
     @Test
     fun winningNumbers() {
-        val winning = Winning(Winning.makeLottoNumbers("1,2,3,4,5,6"), LottoNumber(1))
+        val winning = Winning(WinningNumbers.makeWinningNumbers("1,2,3,4,5,6"), LottoNumber(1))
 
-        winning.numbers.map { it.number } shouldContainInOrder listOf(1, 2, 3, 4, 5, 6)
+        winning.winningNumbers.numbers.map { it.number } shouldContainInOrder listOf(1, 2, 3, 4, 5, 6)
     }
 
     @DisplayName("잘못 입력 된 당첨 번호로 당첨 번호가 생성되지 않는다")
@@ -23,7 +23,7 @@ internal class WinningTest {
     fun wrongWinningNumbers(input: String) {
         assertThatExceptionOfType(IllegalStateException::class.java)
             .isThrownBy {
-                Winning(winningNumbers = Winning.makeLottoNumbers(input), LottoNumber(1))
+                Winning(winningNumbers = WinningNumbers.makeWinningNumbers(input), LottoNumber(1))
             }
     }
 }
